requirements.txt
anyio==3.7.1
attrs==21.4.0
boto==2.49.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cirq==1.1.0
cirq-aqt==1.1.0
cirq-core==1.1.0
cirq-google==1.1.0
cirq-ionq==1.1.0
cirq-pasqal==1.1.0
cirq-rigetti==1.1.0
cirq-web==1.1.0
contourpy==1.1.0
cycler==0.11.0
Deprecated==1.2.14
dill==0.3.7
duet==0.2.8
exceptiongroup==1.1.3
fonttools==4.42.1
future==0.18.3
gekko==1.0.6
git-filter-repo==2.38.0
google-api-core==1.34.0
google-auth==2.22.0
googleapis-common-protos==1.60.0
graphviz==0.20.1
greenlet==3.0.0
grpcio==1.57.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==6.0.1
iniconfig==2.0.0
iso8601==1.1.0
kiwisolver==1.4.5
lark==0.11.3
matplotlib==3.7.2
mpmath==1.3.0
msgpack==1.0.5
networkx==2.8.8
numpy==1.23.5
opentuner==0.8.8
packaging==23.1
pandas==2.0.3
pbr==5.11.1
Pillow==10.0.0
pluggy==1.3.0
ply==3.11
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pydot==1.4.2
PyJWT==2.8.0
pyparsing==3.0.9
pyquil==3.5.4
pytest==7.4.0
pytest-benchmark==4.0.0
python-dateutil==2.8.2
python-rapidjson==1.10
pytz==2023.3
pyzmq==25.1.1
qcs-api-client==0.21.6
qiskit==0.45.0
qiskit-aer==0.13.1
qiskit-terra==0.45.0
requests==2.31.0
retrying==1.3.4
rfc3339==6.2
rfc3986==1.5.0
rpcq==3.11.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
rustworkx==0.13.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.22
stevedore==5.1.0
symengine==0.9.2
sympy==1.12
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
types-Deprecated==1.2.9.3
types-python-dateutil==2.8.19.14
types-retry==0.9.9.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
z3-solver==4.12.2.0
zipp==3.16.2
